+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 - If-Modified-Since HTTP header.
+- If-Unmodified-Since HTTP header.
 
 ### Changed
 - Nothing